# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Plurality API",
    "description": "Description of Plurality API",
    "version": "1.0.0"
  },
  "host": "app.plurality.local:443",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/oauth-facebook/": {
      "get": {
        "tags": [
          "OAuth-Facebook"
        ],
        "description": "",
        "parameters": [
          {
            "name": "sse_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/oauth-facebook/event": {
      "post": {
        "tags": [
          "OAuth-Facebook"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-sse-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/oauth-facebook/info": {
      "get": {
        "tags": [
          "OAuth-Facebook"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "408": {
            "description": "Request Timeout"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/crm/client/": {
      "post": {
        "tags": [
          "Client App"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "img": {
                  "example": "any"
                },
                "streamId": {
                  "example": "any"
                },
                "links": {
                  "example": "any"
                },
                "domains": {
                  "example": "any"
                },
                "incentiveType": {
                  "example": "any"
                },
                "appType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Client App"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-domain",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/crm/client/{id}": {
      "put": {
        "tags": [
          "Client App"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "img": {
                  "example": "any"
                },
                "streamId": {
                  "example": "any"
                },
                "links": {
                  "example": "any"
                },
                "domains": {
                  "example": "any"
                },
                "incentiveType": {
                  "example": "any"
                },
                "appType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}